requirements.txt
openmm>=8.0.0
numpy>=1.20.0
scipy>=1.7.0
matplotlib>=3.5.0
pandas>=1.3.0
PyYAML>=6.0 